requirements.txt
requests==2.26.0
python-dotenv==0.19.0
werkzeug==2.3.7
waitress==2.1.2
flask-cors==4.0.0 